+++ nuxeo-drive-client/nxdrive/engine/workers.py
@@ -6,7 +6,7 @@ from time import sleep, time
 from PyQt4.QtCore import (QCoreApplication, QObject, QThread, pyqtSignal,
                           pyqtSlot)
 
-from nxdrive.engine.activity import Action, IdleAction
+from .activity import Action, IdleAction
 
 log = getLogger(__name__)
 
@@ -227,7 +227,7 @@ class PollWorker(Worker):
         self._check_interval = check_interval
         # Check at start
         self._next_check = 0
-        self._enable = True
+        self.enable = True
         self._metrics = {'last_poll': 0}
 
     def get_metrics(self):
@@ -249,7 +249,7 @@ class PollWorker(Worker):
         self._next_check = 0
 
     def _execute(self):
-        while self._enable:
+        while self.enable:
             self._interact()
             if self.get_next_poll() <= 0:
                 if self._poll():